Fixed compile warning.

+++ src/handle_command.c
@@ -555,7 +555,7 @@ int handle_command(char *line, sm_t *sm_p){
       sm_p->alpcal.timer_length = ftemp;
       printf("CMD: Changed ALP calibration timer length to %.1f seconds\n",sm_p->alpcal.timer_length);
     }else{
-      printf("CMD: ALP calibration timer out of bounds [%d,%d]\n",0,ZERNIKE_ERRORS_LENGTH);
+      printf("CMD: ALP calibration timer out of bounds [%d,%f]\n",0,(double)ZERNIKE_ERRORS_LENGTH);
     }
     return(CMD_NORMAL);
   }
@@ -581,7 +581,7 @@ int handle_command(char *line, sm_t *sm_p){
       sm_p->bmccal.timer_length = ftemp;
       printf("CMD: Changed BMC calibration timer length to %.1f seconds\n",sm_p->bmccal.timer_length);
     }else{
-      printf("CMD: BMC calibration timer out of bounds [%d,%d]\n",0,ZERNIKE_ERRORS_LENGTH);
+      printf("CMD: BMC calibration timer out of bounds [%d,%f]\n",0,(double)ZERNIKE_ERRORS_LENGTH);
     }
     return(CMD_NORMAL);
   }
